+++ src/main/java/dev/jaxydog/content/item/CustomItems.java
@@ -4,6 +4,7 @@ import dev.jaxydog.content.block.CustomBlocks;
 import dev.jaxydog.content.block.custom.DyeableAmethystBlock;
 import dev.jaxydog.content.block.custom.DyeableAmethystClusterBlock;
 import dev.jaxydog.content.block.custom.DyeableBuddingAmethystBlock;
+import dev.jaxydog.content.effect.CustomStatusEffects;
 import dev.jaxydog.content.item.custom.*;
 import dev.jaxydog.datagen.TagGenerator;
 import dev.jaxydog.register.ContentRegistrar;
@@ -124,6 +125,82 @@ public final class CustomItems extends ContentRegistrar {
     public static final CustomItem VOID_ESSENCE = new CustomItem("void_essence",
         new Settings().rarity(Rarity.EPIC),
         CustomItemGroups.STARMONEY_PLAZA
+    );
+    public static final CustomItem DRAGON_SCALE = new CustomItem("dragon_scale",
+        new Settings().rarity(Rarity.EPIC),
+        CustomItemGroups.STARMONEY_PLAZA
+    );
+    public static final RandomEffectItem CLOCK_OF_REGRET = new RandomEffectItem("clock_of_regret",
+        new Settings().maxCount(1).rarity(Rarity.RARE),
+        CustomItemGroups.STARMONEY_PLAZA,
+        0.0005F,
+        StatusEffects.NAUSEA
+    );
+    public static final RandomEffectItem CUP_OF_GRIEF = new RandomEffectItem("cup_of_grief",
+        new Settings().maxCount(1).rarity(Rarity.RARE),
+        CustomItemGroups.STARMONEY_PLAZA,
+        0.0005F,
+        StatusEffects.SLOWNESS
+    );
+    public static final RandomEffectItem SKULL_OF_JOY = new RandomEffectItem("skull_of_joy",
+        new Settings().maxCount(1).rarity(Rarity.RARE),
+        CustomItemGroups.STARMONEY_PLAZA,
+        0.0005F,
+        StatusEffects.REGENERATION
+    );
+    public static final RandomEffectItem BLOB_OF_MALINTENT = new RandomEffectItem("blob_of_malintent",
+        new Settings().maxCount(1).rarity(Rarity.RARE),
+        CustomItemGroups.STARMONEY_PLAZA,
+        0.0005F,
+        CustomStatusEffects.SINISTER
+    );
+    public static final RandomEffectItem COOKIE_OF_RESENTMENT = new RandomEffectItem("cookie_of_resentment",
+        new Settings().maxCount(1).rarity(Rarity.RARE),
+        CustomItemGroups.STARMONEY_PLAZA,
+        0.0005F,
+        StatusEffects.STRENGTH
+    );
+    public static final RandomEffectItem EYE_OF_SURPRISE = new RandomEffectItem("eye_of_surprise",
+        new Settings().maxCount(1).rarity(Rarity.RARE),
+        CustomItemGroups.STARMONEY_PLAZA,
+        0.0005F,
+        StatusEffects.LEVITATION
+    );
+    public static final RandomEffectItem TARGET_OF_PANIC = new RandomEffectItem("target_of_panic",
+        new Settings().maxCount(1).rarity(Rarity.RARE),
+        CustomItemGroups.STARMONEY_PLAZA,
+        0.0005F,
+        StatusEffects.SPEED
+    );
+    public static final RandomEffectItem PHOTO_OF_HOPE = new RandomEffectItem("photo_of_hope",
+        new Settings().maxCount(1).rarity(Rarity.RARE),
+        CustomItemGroups.STARMONEY_PLAZA,
+        0.0005F,
+        StatusEffects.LUCK
+    );
+    public static final RandomEffectItem CROWN_OF_DREAD = new RandomEffectItem("crown_of_dread",
+        new Settings().maxCount(1).rarity(Rarity.RARE),
+        CustomItemGroups.STARMONEY_PLAZA,
+        0.0005F,
+        StatusEffects.DARKNESS
+    );
+    public static final RandomEffectItem EGG_OF_GREED = new RandomEffectItem("egg_of_greed",
+        new Settings().maxCount(1).rarity(Rarity.RARE),
+        CustomItemGroups.STARMONEY_PLAZA,
+        0.0005F,
+        StatusEffects.HERO_OF_THE_VILLAGE
+    );
+    public static final RandomEffectItem FLOWER_OF_SUFFERING = new RandomEffectItem("flower_of_suffering",
+        new Settings().maxCount(1).rarity(Rarity.RARE),
+        CustomItemGroups.STARMONEY_PLAZA,
+        0.0005F,
+        StatusEffects.WITHER
+    );
+    public static final RandomEffectItem BULB_OF_REJECTION = new RandomEffectItem("bulb_of_rejection",
+        new Settings().maxCount(1).rarity(Rarity.RARE),
+        CustomItemGroups.STARMONEY_PLAZA,
+        0.0005F,
+        StatusEffects.INVISIBILITY
     );
 
     @Override